// server/src/types/FinessPharmacy.ts
export interface FinessPharmacy {
    rs: string;
    rslongue: string;
    numvoie: string;
    typvoie: string;
    voie: string;
    compvoie: string;
    lieuditbp: string;
    commune: string;
    departement: string;
    libdepartement: string;
    telephone: string;
    coordxet: number;
    coordyet: number;
}